Arrays and Hashings
Solution to Leet Code Problem - Valid Anagram - Checking  the two input (strings) length, if it is different, returning false. Otherwise,  creating a dictionary, iterating throught the string1, storing the current char in the dictionary as a key, and value 1. If the dict already contains de key, just add 1 to current key. Then, iterating throught the string2, checking it the dict doesn't contains  the current key, if so, returning false, otherwise, diminish 1 to the dict current key. In the last iteration checking if the dict contais value different than 0, if so, returning false, otherwise, return true. Solution validated with Unit Tests.